Now shows seasons and episodes on the other List box

+++ TVshowTracker/MainForm.cs
@@ -14,6 +14,8 @@ namespace TVshowTracker
 {
     public partial class MainForm : Form
     {
+        bool IsShowingEpisodes = false;
+
         public MainForm()
         {
             InitializeComponent();
@@ -22,7 +24,7 @@ namespace TVshowTracker
 
         private void Form1_Load(object sender, EventArgs e)
         {
-            //MaximizeBox = false;
+            MaximizeBox = false;
             ShowTVshows(); // loads show names to TVshowBox
         }
 
@@ -33,16 +35,51 @@ namespace TVshowTracker
             {
                 TVShowBox.Items.Add(string.Format("{0} \n", item.TVname));
             }
-
-
         }
+
+        private void ShowSeasons(int ShowIndex)
+        {
+            if (TVshowManager.TVshows[ShowIndex].Seasons.Count == 0)
+            {
+                EpisodeBox.Items.Add("No seasons available");
+                return;
+            }
+            TVshowManager.TVshows[ShowIndex].SortSeasons();
+            foreach (var item in TVshowManager.TVshows[ShowIndex].Seasons)
+            {
+                EpisodeBox.Items.Add(string.Format("{0} \n", item.SeasonName));
+            }
+            IsShowingEpisodes = false;
+        }
+
+        private void ShowEpisodes(int EpisodeIndex)
+        {
+            if (TVshowManager.TVshows[TVShowBox.SelectedIndex].Seasons[EpisodeIndex].Episodes.Count == 0)
+            {
+                EpisodeBox.Items.Add("No episodes available");
+                return;
+            }
+            EpisodeBox.Items.Clear();
+            foreach (var item in TVshowManager.TVshows[TVShowBox.SelectedIndex].Seasons[EpisodeIndex].Episodes)
+            {
+                EpisodeBox.Items.Add(string.Format("{0} \n", item.EpisodeName));
+            }
+            IsShowingEpisodes = true;
+        }
+
         private void button5_Click(object sender, EventArgs e)
         {
         }
 
         private void TVshows_SelectedIndexChanged(object sender, EventArgs e)
         {
-            
+            int index = TVShowBox.SelectedIndex;
+            if (index < 0)
+            {
+                return;
+            }
+            EpisodeBox.Items.Clear();
+            ShowSeasons(index);
         }
 
         private void button1_Click(object sender, EventArgs e)
@@ -70,11 +107,13 @@ namespace TVshowTracker
 
         private void RemoveShow_Click(object sender, EventArgs e)
         {
-            File.Delete(TVshowManager.TVshows[TVShowBox.SelectedIndex].FilePath); // Deletes show's file in "TV shows" folder
-            TVshowManager.TVshows.RemoveAt(TVShowBox.SelectedIndex); // Removes selected item in listbox from TvshowManager
-            TVShowBox.Items.Clear(); // Clears listbox
-            ShowTVshows(); // Writes updates show list
-            
+            if (TVShowBox.SelectedIndex >= 0)
+            {
+                File.Delete(TVshowManager.TVshows[TVShowBox.SelectedIndex].FilePath); // Deletes show's file in "TV shows" folder
+                TVshowManager.TVshows.RemoveAt(TVShowBox.SelectedIndex); // Removes selected item in listbox from TvshowManager
+                TVShowBox.Items.Clear(); // Clears listbox
+                ShowTVshows(); // Writes updates show list
+            }        
         }
 
         private void MarkedAsWatched_Click(object sender, EventArgs e)
@@ -85,6 +124,18 @@ namespace TVshowTracker
         private void RemoveEpisode_Click(object sender, EventArgs e)
         {
 
+        }
+
+        private void EpisodeBox_SelectedIndexChanged(object sender, EventArgs e)
+        {
+            int index = EpisodeBox.SelectedIndex;
+            if (index < 0 || IsShowingEpisodes)
+            {
+                return;
+            }
+
+            ShowEpisodes(index);
+          
         }
     }
 }
